+++ src/main/java/com/nha/java/learning/phoneshop/exception/ResourceNotFoundException.java
@@ -4,7 +4,7 @@ import org.springframework.http.HttpStatus;
 
 public class ResourceNotFoundException extends ApiException{
 
-	public ResourceNotFoundException(String resourceName, Integer id) {
+	public ResourceNotFoundException(String resourceName, Long id) {
 		super(HttpStatus.NOT_FOUND, "Brand with id = %d not found".formatted(id));
 		
 	}
