+++ src/main/java/com/nha/java/learning/phoneshop/repository/BrandRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.nha.java.learning.phoneshop.entity.Brand;
 
 @Repository
-public interface BrandRepository extends JpaRepository<Brand, Integer>{
+public interface BrandRepository extends JpaRepository<Brand, Long>{
 
 }
