+++ src/main/java/com/nha/java/learning/phoneshop/service/imp/BranServiceImp.java
@@ -21,17 +21,16 @@ public class BranServiceImp implements BrandService{
 	}
 	
 	@Override
-	public Brand getById(Integer id) {
+	public Brand getById(Long id) {
 		return brandRepository.findById(id)
 //				.orElseThrow(()-> new HttpClientErrorException(HttpStatus.NOT_FOUND, "Brand with id = %d not found".formatted(id)));
 				.orElseThrow(()-> new ResourceNotFoundException("Brand", id));
 	}
 
 	@Override
-	public Brand update(Integer id, Brand brandUpdate) {
+	public Brand update(Long id, Brand brandUpdate) {
 		Brand brand = getById(id);
 		brand.setName(brandUpdate.getName());
 		return brandRepository.save(brand);
 	}
-	
 }
